// src/types/supabase.ts
export interface definitions {
  name_changes: {
    /**
     * Format: bigint
     * @description Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    /**
     * Format: timestamp with time zone
     * @default now()
     */
    created_at?: string;
    /** Format: character varying */
    name?: string;
    /** Format: character varying */
    public_key?: string;
    /** Format: character varying */
    signature?: string;
    /** Format: character varying */
    mint_address?: string;
  };
}